Google Cloud Function/PubSub_to_BigQuery/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery~=3.3.0
